requirements.txt
cherrypy
opencv-python-headless
redis
paho-mqtt
b2sdk
boto3
rq
pyyaml
apscheduler
imageio